updating propulsor drag methods for unique propellers

+++ trunk/SUAVE/Methods/Aerodynamics/Common/Fidelity_Zero/Drag/parasite_drag_propulsor.py
@@ -73,7 +73,7 @@ def parasite_drag_propulsor(state,settings,geometry):
     cf_prop, k_comp, k_reyn = compressible_turbulent_flat_plate(Re_prop,Mc,Tc)
     
     ## form factor according to Raymer equation (pg 283 of Aircraft Design: A Conceptual Approach)
-    k_prop = 1 + 0.35 / (float(l_prop)/float(d_prop))  
+    k_prop = 1 + 0.35 / (l_prop/d_prop).astype(float)
     
    
     # find the final result
